\name{generateCompartmentProposal}
\alias{generateCompartmentProposal}
\title{Helper function to find starting values for compartments.}
\description{
    The \code{generateCompartmentProposal} function provides a 
    crude guess at the initial values of the unknown compartments
    based on I_star. 
}



