\name{plotTwoCompartments}
\alias{plotTwoCompartments}
\title{Helper function to visualize compartment data.}
\description{
    The \code{plotTwoCompartments} function accepts two matrix like objects 
    and produces image plots of each with lablels appropriate to comparment 
    matrices and on the same scale. 
}



