\name{spatialSEIRModel}
\alias{spatialSEIRModel}
\title{R function which creates a spatialSEIRInterface object}
\description{
    The spatialSEIRModel function currently provides the API for creating
    spatialSEIRModel objects. This will change in the future in favor 
    of more user friendly constructions. 

}



